# specs/001-wiki-platform/contracts/api.yaml
openapi: 3.0.3
info:
  title: Wiki Platform API
  description: API for the wiki platform with article management, user profiles, and social features
  version: 1.0.0
  contact:
    name: Wiki Platform Support
    email: support@wikiplatform.example.com
servers:
  - url: https://api.wikipatform.example.com/v1
    description: Production server
  - url: https://staging-api.wikipatform.example.com/v1
    description: Staging server
paths:
  /users:
    post:
      summary: Register a new user
      description: Creates a new user account with the provided details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List users
      description: Retrieve a paginated list of users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    get:
      summary: Get user profile
      description: Retrieve a user's public profile information
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update user profile
      description: Update the authenticated user's profile information
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /articles:
    post:
      summary: Create article
      description: Create a new article draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List articles
      description: Retrieve a paginated list of articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArticleResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /comments:
    post:
      summary: Create comment
      description: Create a new comment on an article
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /connections:
    post:
      summary: Create connection
      description: Send a connection request to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionCreate'
      responses:
        '201':
          description: Connection request sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Connection already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /search:
    get:
      summary: Search articles
      description: Search for articles by title, content, or tags
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArticleResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  total:
                    type: integer
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegistration:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          maxLength: 50
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        contributionCount:
          type: integer
        isFollowing:
          type: boolean
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        role:
          type: string
          enum: [user, moderator, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        contributionCount:
          type: integer
...
    ArticleCreate:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
          minLength: 10
        summary:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [draft, published]
          default: draft
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagNames:
          type: array
          items:
            type: string
...
    ArticleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        summary:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        author:
          $ref: '#/components/schemas/UserResponse'
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        viewCount:
          type: integer
        isFeatured:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        parentId:
          type: string
          format: uuid
        articleCount:
          type: integer
    TagResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        usageCount:
          type: integer
...
    CommentCreate:
      type: object
      required:
        - articleId
        - content
      properties:
        articleId:
          type: string
          format: uuid
        content:
          type: string
          minLength: 1
          maxLength: 10000
        parentId:
          type: string
          format: uuid
...
    CommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        author:
          $ref: '#/components/schemas/UserResponse'
        articleId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isApproved:
          type: boolean
    ConnectionCreate:
      type: object
      required:
        - userId
        - connectionType
      properties:
        userId:
          type: string
          format: uuid
        connectionType:
          type: string
          enum: [friend, follow]
    ConnectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        follower:
          $ref: '#/components/schemas/UserResponse'
        followed:
          $ref: '#/components/schemas/UserResponse'
        status:
          type: string
          enum: [pending, accepted, blocked]
        connectionType:
          type: string
          enum: [friend, follow]
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object